Return JSON object instead of array of JSON objects + testing
To test, build the project with `vapor build && vapor run` (or run it in Xcode). Make a `GET` request to `http://127.0.0.1:8080/accounts` and the response should be something like:

```json
{
    "accounts": [
        {
            "id": "87FF92A9-E39E-4994-9084-6ACF6D8E3DD3",
            "type": "bank",
            "balance": 5233
        },
        {
            "id": "BCC66A83-456E-432F-92A5-B89F346E340A",
            "type": "investing",
            "balance": 1869
        },
        {
            "id": "E7268789-0F3B-44AD-897C-5B23B2C79331",
            "type": "investing",
            "balance": 4690
        },
        {
            "id": "A34DC73D-0715-4F7C-8E77-4FB1DC677B6A",
            "type": "high interest savings",
            "balance": 1388
        },
        {
            "id": "B974799D-ECE1-466E-A7C9-8594EE5AA126",
            "type": "bank",
            "balance": 9354
        },
        {
            "id": "6F706A72-931F-4AE1-80C8-7E1F2762878E",
            "type": "bank",
            "balance": 3740
        },
        {
            "id": "C8AAFE00-7B1E-40DB-8111-B1BAF0687FBA",
            "type": "high interest savings",
            "balance": 6988
        }
    ]
}
```
